backend/pyproject.toml
[project]
name = "colazione"
version = "0.1.0"
description = "Programma di pianificazione ferroviaria nativa (ARTURO x Trenord)"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [
    { name = "anto", email = "noreply@example.com" },
]
dependencies = [
    # Web framework
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",

    # DB
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.2",
    "greenlet>=3.0",  # async SQLAlchemy

    # Validation / config
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # Auth
    "bcrypt>=4.1",
    "pyjwt>=2.9",

    # Data manipulation
    "openpyxl>=3.1",
    "numbers-parser>=4.10",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5",
    "mypy>=1.10",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/colazione"]

# =====================================================================
# Tooling
# =====================================================================

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
# Conservative defaults: pyflakes (F), pycodestyle E/W, import sorting (I)
select = ["E", "F", "W", "I", "B", "UP"]
ignore = [
    "E501",  # line too long — gestito dal formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.flake8-bugbear]
# FastAPI usa Depends/Query/Path/etc come default args (pattern standard).
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.Query",
    "fastapi.Path",
    "fastapi.Body",
    "fastapi.Header",
    "fastapi.Cookie",
    "fastapi.Form",
    "fastapi.File",
    "fastapi.Security",
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
exclude = ["alembic/versions/.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-q --strict-markers"

[tool.coverage.run]
source = ["src/colazione"]
omit = ["src/colazione/__main__.py"]
